+++ src/pages/profile.css
@@ -1,4 +1,16 @@
-.blog-container{
+css
+
+ .blogs {
+    display: flex;
+    flex-direction: column;
+    padding: 1rem;
+    margin: 10px;
+    background-color: #c9d1d5;
+    border-radius: 30px;
+    box-sizing: border-box;
+    box-shadow: 5px 5px 5px #7e8c9c;
+}
+ .blog-container {
     width: 80vw;
     height: fit-content;
     background-color: #7e8c9c;
@@ -7,40 +19,36 @@
     color: #0c243c;
     font-family: cursive;
 }
-.blog-container textarea{
+ .blog-container textarea {
     width: 100%;
     height: 100%;
     box-sizing: border-box;
-    
-    
 }
-.blog-container fieldset{
+ .blog-container fieldset {
     border: none;
 }
-.blog-container fieldset textarea {
+ .blog-container fieldset textarea {
     display: block;
     width: 80%;
     height: 70%;
     padding: 0.5rem 2rem;
     margin-bottom: 0.5rem;
     border-radius: 10px;
+    font-family: inherit;
 }
-.blog-container fieldset button {
+ .blog-container fieldset button {
     padding: 5px;
     font-family: inherit;
     font-size: 1.5rem;
     border-radius: 5px;
-    transition: 1s;
+    transition: 0.2s;
 }
-
-.blog-container fieldset button:active {
+ .blog-container fieldset button:active {
     box-shadow: inset 2px 2px 2px black;
     background-color: #55c2c3;
     color: #7e8c9c;
 }
-
-
-.my-posts {
+ .my-posts {
     display: flex;
     flex-direction: column;
     padding: 1rem 2rem;
@@ -53,20 +61,44 @@
     border-radius: 30px;
     box-shadow: 5px 5px 5px steelblue;
 }
-
-
-.my-posts h2 {
+ .my-posts h2 {
     font-size: 2rem;
     color: black;
     margin: 0;
     text-align: center;
-    
-    /* margin: 0.5rem; */
 }
-
-.my-posts p {
+ .my-posts p {
     font-size: 1.5rem;
     line-height: 1.2;
     letter-spacing: 1.2;
     font-family: cursive;
+}
+
+ @media screen and (max-width: 768px) {
+    .blogs {
+        margin: 5px;
+    }
+     .blog-container {
+        width: 90vw;
+    }
+     .blog-container fieldset textarea {
+        width: 100%;
+        padding: 0.5rem;
+    }
+     .my-posts {
+        padding: 1rem;
+        margin-top: 20px;
+    }
+}
+ @media screen and (max-width: 480px) {
+    .c-blogs {
+        margin: 2px;
+    }
+     .blog-container {
+        width: 95vw;
+    }
+     .my-posts {
+        padding: 0.5rem;
+        margin-top: 10px;
+    }
 }
